Test: Move GroupBy tests in SimpleQueryTestBase in separate file
