+++ src/EFCore.Specification.Tests/Query/SimpleQueryTestBase.cs
@@ -2091,321 +2091,6 @@ namespace Microsoft.EntityFrameworkCore.Query
         }
 
         [ConditionalFact]
-        public virtual void GroupBy_anonymous()
-        {
-            AssertQuery<Customer>(
-                cs => cs.Select(c => new { c.City, c.CustomerID }).GroupBy(a => a.City),
-                elementSorter: GroupingSorter<string, object>(),
-                elementAsserter: GroupingAsserter<string, dynamic>(d => d.CustomerID));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_anonymous_with_where()
-        {
-            var countries = new[] { "Argentina", "Austria", "Brazil", "France", "Germany", "USA" };
-            AssertQuery<Customer>(
-                cs => cs.Where(c => countries.Contains(c.Country))
-                    .Select(c => new { c.City, c.CustomerID })
-                    .GroupBy(a => a.City),
-                elementSorter: GroupingSorter<string, object>(),
-                elementAsserter: GroupingAsserter<string, dynamic>(d => d.CustomerID));
-        }
-
-        [ConditionalFact(Skip = "Test does not pass. See issue#7160")]
-        public virtual void GroupBy_anonymous_subquery()
-        {
-            AssertQuery<Customer>(cs =>
-                cs.Select(c => new { c.City, c.CustomerID })
-                    .GroupBy(a => from c2 in cs select c2),
-                assertOrder: true);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_nested_order_by_enumerable()
-        {
-            AssertQuery<Customer>(cs =>
-                cs.Select(c => new { c.Country, c.CustomerID })
-                    .OrderBy(a => a.Country)
-                    .GroupBy(a => a.Country)
-                    .Select(g => g.OrderBy(a => a.CustomerID)),
-                assertOrder: true);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_join_default_if_empty_anonymous()
-        {
-            AssertQuery<Order, OrderDetail>(
-                (os, ods) =>
-                    (from order in os
-                     join orderDetail in ods on order.OrderID equals orderDetail.OrderID into orderJoin
-                     from orderDetail in orderJoin.DefaultIfEmpty()
-                     group new
-                     {
-                         orderDetail.ProductID,
-                         orderDetail.Quantity,
-                         orderDetail.UnitPrice
-                     } by new
-                     {
-                         order.OrderID,
-                         order.OrderDate
-                     }).Where(x => x.Key.OrderID == 10248),
-                elementAsserter: GroupingAsserter<dynamic, dynamic>(d => d.ProductID));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_SelectMany()
-        {
-            AssertQuery<Customer>(
-                cs => cs.GroupBy(c => c.City).SelectMany(g => g),
-                entryCount: 91);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_simple()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID),
-                elementSorter: GroupingSorter<string, Order>(),
-                elementAsserter: GroupingAsserter<string, Order>(o => o.OrderID),
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_simple2()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID).Select(g => g),
-                elementSorter: GroupingSorter<string, Order>(),
-                elementAsserter: GroupingAsserter<string, Order>(o => o.OrderID),
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_first()
-        {
-            AssertSingleResult<Order>(
-                os => os.Where(o => o.CustomerID == "ALFKI").GroupBy(o => o.CustomerID).Cast<object>().First(),
-                asserter: GroupingAsserter<string, Order>(o => o.OrderID),
-                entryCount: 6);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Sum()
-        {
-            AssertQueryScalar<Order>(os => os.GroupBy(o => o.CustomerID).Select(g => g.Sum(o => o.OrderID)));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Count()
-        {
-            AssertQueryScalar<Order>(os => os.GroupBy(o => o.CustomerID).Select(g => g.Count()));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_LongCount()
-        {
-            AssertQueryScalar<Order>(os => os.GroupBy(o => o.CustomerID).Select(g => g.LongCount()));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Shadow()
-        {
-            AssertQuery<Employee>(es =>
-                es.Where(e => EF.Property<string>(e, "Title") == "Sales Representative"
-                              && e.EmployeeID == 1)
-                    .GroupBy(e => EF.Property<string>(e, "Title"))
-                    .Select(g => EF.Property<string>(g.First(), "Title")));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Shadow2()
-        {
-            AssertQuery<Employee>(es =>
-                es.Where(e => EF.Property<string>(e, "Title") == "Sales Representative"
-                              && e.EmployeeID == 1)
-                    .GroupBy(e => EF.Property<string>(e, "Title"))
-                    .Select(g => g.First()));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Shadow3()
-        {
-            AssertQuery<Employee>(es =>
-                es.Where(e => e.EmployeeID == 1)
-                    .GroupBy(e => e.EmployeeID)
-                    .Select(g => EF.Property<string>(g.First(), "Title")));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_Sum_Min_Max_Avg()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID).Select(g =>
-                    new
-                    {
-                        Sum = g.Sum(o => o.OrderID),
-                        Min = g.Min(o => o.OrderID),
-                        Max = g.Max(o => o.OrderID),
-                        Avg = g.Average(o => o.OrderID)
-                    }),
-                e => e.Min + " " + e.Max);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_result_selector()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID, (k, g) =>
-                    new
-                    {
-                        Sum = g.Sum(o => o.OrderID),
-                        Min = g.Min(o => o.OrderID),
-                        Max = g.Max(o => o.OrderID),
-                        Avg = g.Average(o => o.OrderID)
-                    }),
-                e => e.Min + " " + e.Max);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_element_selector_sum()
-        {
-            AssertQueryScalar<Order>(
-                os => os.GroupBy(o => o.CustomerID, o => o.OrderID).Select(g => g.Sum()));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_element_selector()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID, o => o.OrderID)
-                    .OrderBy(g => g.Key)
-                    .Select(g => g.OrderBy(o => o)),
-                assertOrder: true,
-                elementAsserter: CollectionAsserter<int>());
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_element_selector2()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID)
-                    .OrderBy(g => g.Key)
-                    .Select(g => g.OrderBy(o => o.OrderID)),
-                assertOrder: true,
-                elementAsserter: CollectionAsserter<Order>());
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_element_selector3()
-        {
-            AssertQuery<Employee>(
-                es => es.GroupBy(e => e.EmployeeID)
-                    .OrderBy(g => g.Key)
-                    .Select(g => g.Select(e => new { Title = EF.Property<string>(e, "Title"), e }).ToList()),
-                assertOrder: true);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_element_selector_sum_max()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID, o => o.OrderID)
-                    .Select(g => new { Sum = g.Sum(), Max = g.Max() }),
-                e => e.Sum + " " + e.Max);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_anonymous_element()
-        {
-            AssertQueryScalar<Order>(os =>
-                os.GroupBy(o => o.CustomerID, o => new { o.OrderID })
-                    .Select(g => g.Sum(x => x.OrderID)));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_two_part_key()
-        {
-            AssertQueryScalar<Order>(os =>
-                os.GroupBy(o => new { o.CustomerID, o.OrderDate })
-                    .Select(g => g.Sum(o => o.OrderID)));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_DateTimeOffset_Property()
-        {
-            AssertQuery<Order>(
-                os => os.Where(o => o.OrderDate.HasValue).GroupBy(o => o.OrderDate.Value.Month),
-                e => ((IGrouping<int, Order>)e).Key,
-                elementAsserter: GroupingAsserter<int, Order>(o => o.OrderID),
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void OrderBy_GroupBy()
-        {
-            AssertQueryScalar<Order>(os =>
-                os.OrderBy(o => o.OrderID)
-                    .GroupBy(o => o.CustomerID)
-                    .Select(g => g.Sum(o => o.OrderID)));
-        }
-
-        [ConditionalFact]
-        public virtual void OrderBy_GroupBy_SelectMany()
-        {
-            AssertQuery<Order>(os =>
-                os.OrderBy(o => o.OrderID)
-                    .GroupBy(o => o.CustomerID)
-                    .SelectMany(g => g),
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void OrderBy_GroupBy_SelectMany_shadow()
-        {
-            AssertQuery<Employee>(es =>
-                es.OrderBy(e => e.EmployeeID)
-                    .GroupBy(e => e.EmployeeID)
-                    .SelectMany(g => g)
-                    .Select(g => EF.Property<string>(g, "Title")));
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_orderby()
-        {
-            AssertQuery<Order>(
-                os => os.OrderBy(o => o.OrderID).GroupBy(o => o.CustomerID).OrderBy(g => g.Key),
-                assertOrder: true,
-                elementAsserter: GroupingAsserter<string, Order>(),
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_orderby_and_anonymous_projection()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID).OrderBy(g => g.Key).Select(g => new { Foo = "Foo", Group = g }),
-                e => GroupingSorter<string, object>()(e.Group),
-                elementAsserter: (e, a) =>
-                    {
-                        Assert.Equal(e.Foo, a.Foo);
-                        IGrouping<string, Order> eGrouping = e.Group;
-                        IGrouping<string, Order> aGrouping = a.Group;
-                        Assert.Equal(eGrouping.OrderBy(p => p.OrderID), aGrouping.OrderBy(p => p.OrderID));
-                    },
-                entryCount: 830);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_with_orderby_take_skip_distinct()
-        {
-            AssertQuery<Order>(
-                os => os.GroupBy(o => o.CustomerID).OrderBy(g => g.Key).Take(5).Skip(3).Distinct(),
-                assertOrder: true,
-                elementAsserter: GroupingAsserter<string, Order>(o => o.OrderID),
-                entryCount: 31);
-        }
-
-        [ConditionalFact]
         public virtual void OrderBy_arithmetic()
         {
             AssertQuery<Employee>(
@@ -2793,7 +2478,7 @@ namespace Microsoft.EntityFrameworkCore.Query
             {
                 var orders
                     = (from o in context.Orders.Take(1)
-                       // ReSharper disable once UseMethodAny.0
+                           // ReSharper disable once UseMethodAny.0
                        where (from od in context.OrderDetails.OrderBy(od => od.OrderID).Take(2)
                               where (from c in context.Set<Customer>()
                                      where c.CustomerID == o.CustomerID
@@ -3746,27 +3431,6 @@ namespace Microsoft.EntityFrameworkCore.Query
                      where customers.Any()
                      select customers).Any()),
                 entryCount: 1);
-        }
-
-        [ConditionalFact]
-        public virtual void GroupBy_join_anonymous()
-        {
-            AssertQuery<Order, OrderDetail>(
-                (os, ods) =>
-                    (from order in os
-                     join orderDetail in ods on order.OrderID equals orderDetail.OrderID into orderJoin
-                     from orderDetail in orderJoin
-                     group new
-                     {
-                         orderDetail.ProductID,
-                         orderDetail.Quantity,
-                         orderDetail.UnitPrice
-                     } by new
-                     {
-                         order.OrderID,
-                         order.OrderDate
-                     }).Where(x => x.Key.OrderID == 10248),
-                elementAsserter: GroupingAsserter<dynamic, dynamic>(d => d.ProductID));
         }
 
         [ConditionalFact]
