+++ test/EFCore.SqlServer.FunctionalTests/Query/SimpleQuerySqlServerTest.GroupBy.cs
@@ -1,0 +1,436 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.EntityFrameworkCore.Query
+{
+    public partial class SimpleQuerySqlServerTest
+    {
+        public override void GroupBy_anonymous()
+        {
+            base.GroupBy_anonymous();
+
+            AssertSql(
+                @"SELECT [c].[City], [c].[CustomerID]
+FROM [Customers] AS [c]
+ORDER BY [c].[City]");
+        }
+
+        public override void GroupBy_anonymous_with_where()
+        {
+            base.GroupBy_anonymous_with_where();
+
+            AssertSql(
+                @"SELECT [c].[City], [c].[CustomerID]
+FROM [Customers] AS [c]
+WHERE [c].[Country] IN (N'Argentina', N'Austria', N'Brazil', N'France', N'Germany', N'USA')
+ORDER BY [c].[City]");
+        }
+
+        public override void GroupBy_anonymous_subquery()
+        {
+            base.GroupBy_anonymous_subquery();
+
+            AssertSql(" ");
+        }
+
+        public override void GroupBy_nested_order_by_enumerable()
+        {
+            base.GroupBy_nested_order_by_enumerable();
+
+            AssertSql(
+                @"SELECT [c0].[Country], [c0].[CustomerID]
+FROM [Customers] AS [c0]
+ORDER BY [c0].[Country]");
+        }
+
+        public override void GroupBy_join_default_if_empty_anonymous()
+        {
+            base.GroupBy_join_default_if_empty_anonymous();
+
+            AssertSql(
+                @"SELECT [order0].[OrderID], [order0].[CustomerID], [order0].[EmployeeID], [order0].[OrderDate], [orderDetail0].[OrderID], [orderDetail0].[ProductID], [orderDetail0].[Discount], [orderDetail0].[Quantity], [orderDetail0].[UnitPrice]
+FROM [Orders] AS [order0]
+LEFT JOIN [Order Details] AS [orderDetail0] ON [order0].[OrderID] = [orderDetail0].[OrderID]
+ORDER BY [order0].[OrderID]");
+        }
+
+        public override void GroupBy_SelectMany()
+        {
+            base.GroupBy_SelectMany();
+
+            AssertSql(
+                @"SELECT [c0].[CustomerID], [c0].[Address], [c0].[City], [c0].[CompanyName], [c0].[ContactName], [c0].[ContactTitle], [c0].[Country], [c0].[Fax], [c0].[Phone], [c0].[PostalCode], [c0].[Region]
+FROM [Customers] AS [c0]
+ORDER BY [c0].[City]");
+        }
+
+        public override void GroupBy_simple()
+        {
+            base.GroupBy_simple();
+
+            AssertSql(
+                @"SELECT [o].[OrderID], [o].[CustomerID], [o].[EmployeeID], [o].[OrderDate]
+FROM [Orders] AS [o]
+ORDER BY [o].[CustomerID]");
+        }
+
+        public override void GroupBy_simple2()
+        {
+            base.GroupBy_simple2();
+
+            AssertSql(
+                @"SELECT [o].[OrderID], [o].[CustomerID], [o].[EmployeeID], [o].[OrderDate]
+FROM [Orders] AS [o]
+ORDER BY [o].[CustomerID]");
+        }
+
+        public override void GroupBy_first()
+        {
+            base.GroupBy_first();
+
+            AssertSql(
+                @"SELECT [o].[OrderID], [o].[CustomerID], [o].[EmployeeID], [o].[OrderDate]
+FROM [Orders] AS [o]
+WHERE [o].[CustomerID] = N'ALFKI'
+ORDER BY [o].[CustomerID]");
+        }
+
+        public override void GroupBy_Sum()
+        {
+            base.GroupBy_Sum();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_Count()
+        {
+            base.GroupBy_Count();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_LongCount()
+        {
+            base.GroupBy_LongCount();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_Shadow()
+        {
+            base.GroupBy_Shadow();
+
+            AssertSql(
+                @"SELECT [e0].[EmployeeID], [e0].[City], [e0].[Country], [e0].[FirstName], [e0].[ReportsTo], [e0].[Title]
+FROM [Employees] AS [e0]
+WHERE ([e0].[Title] = N'Sales Representative') AND ([e0].[EmployeeID] = 1)
+ORDER BY [e0].[Title]");
+        }
+
+        public override void GroupBy_Shadow2()
+        {
+            base.GroupBy_Shadow2();
+
+            AssertSql(
+                @"SELECT [e0].[EmployeeID], [e0].[City], [e0].[Country], [e0].[FirstName], [e0].[ReportsTo], [e0].[Title]
+FROM [Employees] AS [e0]
+WHERE ([e0].[Title] = N'Sales Representative') AND ([e0].[EmployeeID] = 1)
+ORDER BY [e0].[Title]");
+        }
+
+        public override void GroupBy_Shadow3()
+        {
+            base.GroupBy_Shadow3();
+
+            AssertSql(
+                @"SELECT [e0].[EmployeeID], [e0].[City], [e0].[Country], [e0].[FirstName], [e0].[ReportsTo], [e0].[Title]
+FROM [Employees] AS [e0]
+WHERE [e0].[EmployeeID] = 1
+ORDER BY [e0].[EmployeeID]");
+        }
+
+        public override void GroupBy_Sum_Min_Max_Avg()
+        {
+            base.GroupBy_Sum_Min_Max_Avg();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_result_selector()
+        {
+            base.GroupBy_with_result_selector();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_element_selector_sum()
+        {
+            base.GroupBy_with_element_selector_sum();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_element_selector()
+        {
+            base.GroupBy_with_element_selector();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_element_selector2()
+        {
+            base.GroupBy_with_element_selector2();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_element_selector3()
+        {
+            base.GroupBy_with_element_selector3();
+
+            AssertSql(
+                @"SELECT [e0].[EmployeeID], [e0].[City], [e0].[Country], [e0].[FirstName], [e0].[ReportsTo], [e0].[Title]
+FROM [Employees] AS [e0]
+ORDER BY [e0].[EmployeeID]");
+        }
+
+        public override void GroupBy_with_element_selector_sum_max()
+        {
+            base.GroupBy_with_element_selector_sum_max();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_anonymous_element()
+        {
+            base.GroupBy_with_anonymous_element();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_two_part_key()
+        {
+            base.GroupBy_with_two_part_key();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID], [o0].[OrderDate]");
+        }
+
+        public override void GroupBy_DateTimeOffset_Property()
+        {
+            base.GroupBy_DateTimeOffset_Property();
+
+            AssertSql(
+                @"SELECT [o].[OrderID], [o].[CustomerID], [o].[EmployeeID], [o].[OrderDate]
+FROM [Orders] AS [o]
+WHERE [o].[OrderDate] IS NOT NULL
+ORDER BY DATEPART(month, [o].[OrderDate])");
+        }
+
+        public override void OrderBy_GroupBy()
+        {
+            base.OrderBy_GroupBy();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID], [o0].[OrderID]");
+        }
+
+        public override void OrderBy_GroupBy_SelectMany()
+        {
+            base.OrderBy_GroupBy_SelectMany();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID], [o0].[OrderID]");
+        }
+
+        public override void OrderBy_GroupBy_SelectMany_shadow()
+        {
+            base.OrderBy_GroupBy_SelectMany_shadow();
+
+            AssertSql(
+                @"SELECT [e0].[EmployeeID], [e0].[City], [e0].[Country], [e0].[FirstName], [e0].[ReportsTo], [e0].[Title]
+FROM [Employees] AS [e0]
+ORDER BY [e0].[EmployeeID]");
+        }
+
+        public override void GroupBy_with_orderby()
+        {
+            base.GroupBy_with_orderby();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID], [o0].[OrderID]");
+        }
+
+        public override void GroupBy_with_orderby_and_anonymous_projection()
+        {
+            base.GroupBy_with_orderby_and_anonymous_projection();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_with_orderby_take_skip_distinct()
+        {
+            base.GroupBy_with_orderby_take_skip_distinct();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_join_anonymous()
+        {
+            base.GroupBy_join_anonymous();
+
+            AssertSql(
+                @"SELECT [order0].[OrderID], [order0].[CustomerID], [order0].[EmployeeID], [order0].[OrderDate], [orderDetail0].[OrderID], [orderDetail0].[ProductID], [orderDetail0].[Discount], [orderDetail0].[Quantity], [orderDetail0].[UnitPrice]
+FROM [Orders] AS [order0]
+LEFT JOIN [Order Details] AS [orderDetail0] ON [order0].[OrderID] = [orderDetail0].[OrderID]
+ORDER BY [order0].[OrderID]");
+        }
+
+        public override void Select_GroupBy_All()
+        {
+            base.Select_GroupBy_All();
+
+            AssertSql(
+                @"SELECT [o].[OrderID] AS [Order], [o].[CustomerID] AS [Customer]
+FROM [Orders] AS [o]
+ORDER BY [o].[CustomerID]");
+        }
+
+        public override void Select_GroupBy()
+        {
+            base.Select_GroupBy();
+
+            AssertSql(
+                @"SELECT [o].[OrderID] AS [Order], [o].[CustomerID] AS [Customer]
+FROM [Orders] AS [o]
+ORDER BY [o].[CustomerID]");
+        }
+
+        public override void Select_GroupBy_SelectMany()
+        {
+            base.Select_GroupBy_SelectMany();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID] AS [Order], [o0].[CustomerID] AS [Customer]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[OrderID]");
+        }
+
+        public override void Distinct_GroupBy()
+        {
+            base.Distinct_GroupBy();
+
+            AssertSql(
+                @"SELECT DISTINCT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void GroupBy_Distinct()
+        {
+            base.GroupBy_Distinct();
+
+            AssertSql(
+                @"SELECT [o0].[OrderID], [o0].[CustomerID], [o0].[EmployeeID], [o0].[OrderDate]
+FROM [Orders] AS [o0]
+ORDER BY [o0].[CustomerID]");
+        }
+
+        public override void Union_simple_groupby()
+        {
+            base.Union_simple_groupby();
+
+            AssertSql(" ");
+        }
+
+        public override void Select_nested_collection_with_groupby()
+        {
+            base.Select_nested_collection_with_groupby();
+
+            AssertSql(
+                @"SELECT (
+    SELECT CASE
+        WHEN EXISTS (
+            SELECT 1
+            FROM [Orders] AS [o0]
+            WHERE [c].[CustomerID] = [o0].[CustomerID])
+        THEN CAST(1 AS BIT) ELSE CAST(0 AS BIT)
+    END
+), [c].[CustomerID]
+FROM [Customers] AS [c]
+WHERE [c].[CustomerID] LIKE N'A' + N'%' AND (LEFT([c].[CustomerID], LEN(N'A')) = N'A')",
+                //
+                @"@_outer_CustomerID1='ALFKI' (Size = 5)
+
+SELECT [o2].[OrderID], [o2].[CustomerID], [o2].[EmployeeID], [o2].[OrderDate]
+FROM [Orders] AS [o2]
+WHERE @_outer_CustomerID1 = [o2].[CustomerID]
+ORDER BY [o2].[OrderID]",
+                //
+                @"@_outer_CustomerID1='ANATR' (Size = 5)
+
+SELECT [o2].[OrderID], [o2].[CustomerID], [o2].[EmployeeID], [o2].[OrderDate]
+FROM [Orders] AS [o2]
+WHERE @_outer_CustomerID1 = [o2].[CustomerID]
+ORDER BY [o2].[OrderID]",
+                //
+                @"@_outer_CustomerID1='ANTON' (Size = 5)
+
+SELECT [o2].[OrderID], [o2].[CustomerID], [o2].[EmployeeID], [o2].[OrderDate]
+FROM [Orders] AS [o2]
+WHERE @_outer_CustomerID1 = [o2].[CustomerID]
+ORDER BY [o2].[OrderID]",
+                //
+                @"@_outer_CustomerID1='AROUT' (Size = 5)
+
+SELECT [o2].[OrderID], [o2].[CustomerID], [o2].[EmployeeID], [o2].[OrderDate]
+FROM [Orders] AS [o2]
+WHERE @_outer_CustomerID1 = [o2].[CustomerID]
+ORDER BY [o2].[OrderID]");
+        }
+    }
+}
